test(helpdesk): merge chat message fixtures and unskip link tests
