+++ tests/e2e/helpdesk/chats/links/test_sync_links.py
@@ -8,21 +8,18 @@ from mpt_api_client.resources.helpdesk.chat_links import ChatLink
 pytestmark = [pytest.mark.flaky]
 
 
-@pytest.mark.skip(reason="Unskip after MPT-19124 completed")
-def test_list_chat_links(chat_links_service):
-    result = chat_links_service.fetch_page(limit=1)
+def test_list_chat_links(created_chat_link_service):
+    result = created_chat_link_service.fetch_page(limit=1)
 
     assert len(result) > 0
     assert all(isinstance(link, ChatLink) for link in result)
 
 
-@pytest.mark.skip(reason="Unskip after MPT-19124 completed")  # noqa: AAA01
-def test_create_chat_link(created_chat_link, chat_link_data):
+def test_create_chat_link(created_chat_link, chat_link_data):  # noqa: AAA01
     assert created_chat_link.id is not None
     assert created_chat_link.to_dict().get("uri") == chat_link_data["uri"]
 
 
-@pytest.mark.skip(reason="Unskip after MPT-19124 completed")
 def test_update_chat_link_name(chat_links_service, created_chat_link, short_uuid):
     new_name = f"e2e updated link - {short_uuid}"
 
@@ -32,7 +29,6 @@ def test_update_chat_link_name(chat_links_service, created_chat_link, short_uuid
     assert result.to_dict().get("name") == new_name
 
 
-@pytest.mark.skip(reason="Unskip after MPT-19124 completed")
 def test_delete_chat_link(chat_links_service, created_chat_link):
     result = created_chat_link
